ExoPlayer2: Narration tracks break audio track switching
Setting a Audio track to a language code of "nar", or other
made up language codes (as against defined in ISO 639-1/639-2,
will result in attempts to change audio tracks to simply fail.

By not setting the language (As we have no idea if this is a english
or frech etc narration), we'll show this as a Unknown Language track,
but it will at least work.