Remove use of versionName in archivesBaseName
Android Studio has a hard time tracking the APKs to use when the name
changes. Long story short, it caches the name when a Gradle Sync is
performed, but generates a different name whenever any git commits
or branch changes etc happens.. The end result is you build, and it'll
upload an old APK to your device until the next Gradle Sync.