Half fix channel sorting on Sony TVs
It's impossible for us to fully fix this, as soon as a new channel is added or
updated, it will be added at the end no matter what channel number it is.

This only proper fix for this can come from Sony.

In the meantime, Sony users must uninstall, reinstall, and resetup the app in
order to resort the channels

Fixes #92